# goforms/contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GoFormX API",
    "version": "1.2.0",
    "license": {
      "name": "GNU Affero General Public License v3.0 or later",
      "identifier": "AGPL-3.0-or-later"
    },
    "description": "The control plane manages forms and immutable schema versions. The public\ndata plane exposes published schemas and accepts validated submissions.\nSchema and submission JSON numbers preserve exact numeric values. Clients\nmust not round numbers through binary floating point before sending them or\nafter reading them. Numeric tokens are limited to 4096 bytes, explicit\nexponents to -1024 through 1024, and expanded values to 1024 integer and\n1024 fractional decimal places. Excessive numeric representations return\n400 rather than being rounded. JSONB may normalize numeric notation;\nnumerically equivalent retries remain idempotent.\nEvery requestBody operation requires exactly its documented media type.\nContent-Type may omit parameters or include only `charset=utf-8`\n(case-insensitive); missing headers, other media types, charsets, or\nparameters return 415 `unsupported_media_type`. Request documents must be\none JSON object. Duplicate object member names are rejected before decoding\nat every nesting depth, including escape-equivalent names such as `a` and\n`\\u0061`; malformed or trailing JSON returns 400 `invalid_request`.\n"
  },
...
  "servers": [
    {
      "url": "https://api.goformx.com",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/v1/service-tokens": {
      "get": {
        "summary": "List service-token metadata for the authenticated organization",
        "operationId": "listServiceTokens",
        "tags": [
          "Control plane"
        ],
        "security": [
          {
            "serviceToken": []
          },
          {
            "firstPartyAssertion": []
          }
        ],
        "x-goformx-required-scopes": [
          "tokens:read"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "Bounded token metadata; hashes and plaintext credentials are never returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ServiceToken"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "limit"
                      ],
                      "additionalProperties": false,
                      "properties": {
                        "limit": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 100
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "summary": "Issue an organization service token with one-time plaintext reveal",
        "operationId": "createServiceToken",
        "tags": [
          "Control plane"
        ],
        "security": [
          {
            "serviceToken": []
          },
          {
            "firstPartyAssertion": []
          }
        ],
        "x-goformx-required-scopes": [
          "tokens:write"
        ],
        "description": "The caller may delegate only scopes already present in its authenticated principal. Issuance commits atomically with an append-only caller-attributed audit record; an audit failure returns 503 without issuing or revealing a token. This operation is intentionally not retryable because GoFormX cannot recover a stored token secret.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceToken"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token issued; the response is Cache-Control no-store and plaintext is never recoverable",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedServiceTokenEnvelope"
                }
              }
            }
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/v1/service-tokens/{tokenId}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/TokenId"
        }
      ],
      "delete": {
        "summary": "Revoke one service token owned by the authenticated organization",
        "operationId": "revokeServiceToken",
        "description": "Revocation and its caller-attributed audit record commit atomically. Repeated revocation of an owned token succeeds without changing its original revocation time or duplicating the mutation audit. Audit failure returns 503 without revoking the token.",
        "tags": [
          "Control plane"
        ],
        "security": [
          {
            "serviceToken": []
          },
          {
            "firstPartyAssertion": []
          }
        ],
        "x-goformx-required-scopes": [
          "tokens:write"
        ],
        "responses": {
          "204": {
            "description": "Token revoked"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "serviceToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "gfst_ opaque token",
        "description": "External integration credential stored server-side, returned once, and verified by hash, organization owner, lifecycle, and required operation scope. Browsers must not receive it.",
        "x-goformx-credential-class": "external-service-token"
      },
      "firstPartyAssertion": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT (EdDSA)",
        "description": "Single-use, at-most-60-second assertion minted by the Waaseyaa server after session authentication and organization membership resolution. Never delivered to a browser and never accepted as an external service token.",
        "x-goformx-credential-class": "first-party-assertion",
        "x-goformx-jwt-profile": "gofx-fpa-v1",
        "x-goformx-jwks-uri": "https://goformx.com/.well-known/goformx-control-plane-jwks.json"
      }
    },
    "parameters": {
...
      "TokenId": {
        "name": "tokenId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[A-Za-z0-9_-]{16}$"
        }
      },
...
      "PageLimit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum resources returned in this bounded collection.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 25
        }
      },
...
    },
    "schemas": {
...
      "CreateServiceToken": {
        "type": "object",
        "required": [
          "name",
          "scopes"
        ],
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "scopes": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "enum": [
                "forms:read",
                "forms:write",
                "forms:publish",
                "submissions:read",
                "tokens:read",
                "tokens:write",
                "webhooks:read",
                "webhooks:write"
              ]
            }
          },
          "expiresInSeconds": {
            "type": "integer",
            "minimum": 300,
            "maximum": 31536000,
            "default": 2592000
          }
        }
      },
      "ServiceToken": {
        "type": "object",
        "required": [
          "id",
          "name",
          "organizationId",
          "scopes",
          "status",
          "createdAt",
          "expiresAt"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{16}$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "scopes": {
            "type": "array",
            "minItems": 1,
            "uniqueItems": true,
            "items": {
              "type": "string",
              "enum": [
                "forms:read",
                "forms:write",
                "forms:publish",
                "submissions:read",
                "tokens:read",
                "tokens:write",
                "webhooks:read",
                "webhooks:write"
              ]
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "expired",
              "revoked"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastUsedAt": {
            "type": "string",
            "format": "date-time"
          },
          "revokedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreatedServiceTokenEnvelope": {
        "type": "object",
        "required": [
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "token",
              "metadata"
            ],
            "additionalProperties": false,
            "properties": {
              "token": {
                "type": "string",
                "pattern": "^gfst_[A-Za-z0-9_-]{43}$",
                "writeOnly": true
              },
              "metadata": {
                "$ref": "#/components/schemas/ServiceToken"
              }
            }
          }
        }
      },
...
      "FieldError": {
        "type": "object",
        "required": [
          "pointer",
          "code",
          "message"
        ],
        "properties": {
          "pointer": {
            "type": "string",
            "description": "JSON Pointer into the request document"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "fields": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                }
              }
            }
          }
        }
      }
    },
    "responses": {
      "Error": {
        "description": "Stable error envelope; status is determined by the HTTP response code",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "UnsupportedMediaType": {
        "description": "Content-Type is missing or does not match the operation's documented JSON media type and optional UTF-8 charset policy",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "unsupported_media_type",
                "message": "Content-Type must match the operation's documented JSON media type; only an optional UTF-8 charset parameter is accepted.",
                "requestId": "req_01K4Q9X2JXQG67A4WZ8X9CEQ4G"
              }
            }
          }
        }
      },
...
    }
  },
...
}